requirements.txt
Flask==3.0.1
feedparser==6.0.11
requests==2.31.0
APScheduler==3.10.4
python-dotenv==1.0.0
openai==1.8.0
pillow==10.2.0 
RPi.GPIO==0.7.1